requirements.txt
opencv-python~=4.6.0.66
numpy~=1.21.6
matplotlib~=3.5.2
pillow~=9.2.0
pywin32~=304
pyautogui~=0.9.53
Desktopmagic